Add integration with the module workbench_moderation.
Several importants changes:
The architecture of the hooks implemented by srf is changed,
now you should define to functions in the hook implementation,
one for alter the widget and other for alter the validation of the widget.
Added a global validation function that will call the callback_validations
functions of the hook if is really necesary the validation.
Major changes in the srf_form_field_ui_field_edit_form_alter to integrate
the workbench status.

Refs:
====
Drupal R&D:
Story #72390

Dauphine:
Tâche / Task #72369
Scénario / Story #66068